LandDetector: Merged fixedwing and multicopter into one module handling both algorithms
