--- src/modules/fw_land_detector/module.mk
@@ -1,10 +1,0 @@
-#
-# Fixedwing land detector
-#
-
-MODULE_COMMAND  = fw_land_detector
-
-SRCS        = fw_land_detector_main.cpp \
-              FixedwingLandDetector.cpp
-
-EXTRACXXFLAGS   = -Weffc++
