+++ src/modules/land_detector/module.mk
@@ -1,0 +1,12 @@
+#
+# Land detector
+#
+
+MODULE_COMMAND	= land_detector
+
+SRCS		= land_detector_main.cpp \
+              LandDetector.cpp \
+			  MulticopterLandDetector.cpp \
+              FixedwingLandDetector.cpp
+
+EXTRACXXFLAGS   = -Weffc++ -Os
