--- src/modules/mc_land_detector/module.mk
@@ -1,10 +1,0 @@
-#
-# Multirotor land detector
-#
-
-MODULE_COMMAND	= mc_land_detector
-
-SRCS		= mc_land_detector_main.cpp \
-			  MulticopterLandDetector.cpp
-
-EXTRACXXFLAGS   = -Weffc++
